requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.9
langchain-openai==0.0.2
langgraph==0.0.20
openai>=1.6.1
streamlit==1.29.0
requests==2.31.0
pydantic>=2.8
langsmith>=0.1.115
